feat: Adiciona DTOs para as entidades Produto e Usuário
* Criados os DTOs ProdutoCreateDTO, ProdutoUpdateDTO e ProdutoResponseDTO para gerenciamento de dados de produtos.
* Introduzidos UsuarioCreateDTO e UsuarioUpdateDTO para gerenciamento de usuários.
* Implementado AlterarSenhaDTO para funcionalidade de alteração de senha.

refactor: Atualiza entidades e mapeadores

* Refatoradas as entidades MovimentacaoEstoque e Produto para utilizar convenções de nomenclatura e relacionamentos adequados.
* Atualizados os mapeadores para suportar os novos DTOs e garantir o mapeamento correto dos dados.
* Aprimorada a entidade Produto com versionamento para controle de concorrência otimista (optimistic locking).

fix: Melhora métodos dos repositórios para tratamento de dados

* Modificados os repositórios MovimentacaoRepository e ProdutoRepository para utilizar paginação e Entity Graphs, proporcionando uma recuperação de dados mais eficiente.
* Adicionado mecanismo de bloqueio no ProdutoRepository para evitar modificações concorrentes.

chore: Configura propriedades da aplicação para diferentes ambientes

* Criados os arquivos application-dev.properties e application-prod.properties para configurações específicas de cada ambiente.
* Atualizado o application.properties para simplificar o gerenciamento das configurações.

db: Cria migração inicial do esquema do banco de dados

* Adicionado o arquivo V1__create_initial_schema.sql para configurar a estrutura inicial do banco de dados, incluindo as tabelas e restrições necessárias.

+++ src/main/java/br/senai/saepveterinaria/dto/produto/ProdutoResponseDTO.java
@@ -3,6 +3,7 @@ package br.senai.saepveterinaria.dto.produto;
 import br.senai.saepveterinaria.enums.TipoProduto;
 import br.senai.saepveterinaria.enums.UnidadeMedida;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ProdutoResponseDTO(
@@ -14,8 +15,8 @@ public record ProdutoResponseDTO(
         Integer quantidadeEstoque,
         Integer estoqueMinimo,
         LocalDate validade,
-        Double pesoKg,
-        Double dosagem,
+        BigDecimal pesoKg,
+        BigDecimal dosagem,
         UnidadeMedida unidadeMedida
 
 ) {
